feat(react): updates input select
