--- packages/react/src/components/form/components/InputSelectSearch/InputSelectSearch.tsx
@@ -1,186 +1,0 @@
-/* eslint-disable react/no-array-index-key */
-/* eslint-disable no-nested-ternary */
-/* eslint-disable @typescript-eslint/no-explicit-any */
-/* eslint-disable jsx-a11y/no-noninteractive-tabindex */
-import * as React from 'react';
-import { useCombobox, UseComboboxStateChange } from 'downshift';
-import { IconChevronStepper } from '@aksara-ui/icons';
-
-import { Box, Stack } from '../../../../layout';
-import { FormLabel, Card, InputText } from '../../..';
-import { useComponentStyles } from '../../../../system';
-
-export interface InputSelectSearchProps<T extends { value: string }> {
-  /** The input select label */
-  label?: string;
-  /** Placeholder text for select label */
-  placeholder?: string;
-  /** Label items. */
-  items: T[];
-  /** Selected item. */
-  selectedItem?: T | null;
-
-  initialSelectedItem?: T | null;
-  /** If the item list is an object/shape, use this to map it into string. */
-  itemToString?: (item: T | null) => string;
-  /** The change handler for the select. */
-  handleSelectedItemChange?: (changes: UseComboboxStateChange<T>) => void;
-  /** If the item list is an object/shape, use this to map a custom element to render on the UI. */
-  itemRenderer?: (item: T) => React.ReactNode;
-  /** Name of the field form */
-  name?: string;
-  /** open list when onfocus */
-  openOnFocus?: boolean;
-  /** Logic on focus */
-  onFocus?: () => void;
-  /** Logic on blue */
-  onBlur?: () => void;
-
-  disabled?: boolean;
-
-  size?: 'md' | 'lg';
-
-  errors?: boolean;
-
-  width?: string | number;
-  /** Max height for list box */
-  maxHeight?: string | number;
-}
-
-/** Base wrapper for dropdown selector element using Downshift.js */
-function InputSelect<T extends { value: string }>({
-  label,
-  placeholder = 'Select an item',
-  items,
-  selectedItem,
-  itemToString,
-  handleSelectedItemChange,
-  itemRenderer,
-  initialSelectedItem,
-  onBlur,
-  onFocus,
-  disabled,
-  openOnFocus = false,
-  errors,
-  size = 'md',
-  width = '100%',
-  maxHeight,
-}: InputSelectSearchProps<T>) {
-  const [inputItems, setInputItems] = React.useState(items);
-  const {
-    isOpen,
-    getLabelProps,
-    getMenuProps,
-    highlightedIndex,
-    getItemProps,
-    getInputProps,
-    getComboboxProps,
-    openMenu,
-  } = useCombobox<T>({
-    items: inputItems,
-    itemToString,
-    selectedItem,
-    initialSelectedItem,
-    onSelectedItemChange: handleSelectedItemChange,
-    onInputValueChange: ({ inputValue }) => {
-      if (inputValue) {
-        setInputItems(items.filter(item => item.value.toLowerCase().startsWith(inputValue.toLowerCase())));
-      }
-    },
-  });
-
-  const styles = useComponentStyles('inputText', { size, variant: errors ? 'error' : isOpen ? 'active' : 'default' });
-
-  return (
-    <Box width={width}>
-      <Stack spacing="xxs" display="block" position="relative">
-        {label && (
-          <FormLabel display="block" {...getLabelProps()}>
-            {label}
-          </FormLabel>
-        )}
-        <Box display="flex" position="relative" alignItems="center" {...getComboboxProps()}>
-          <InputText
-            disabled={disabled}
-            placeholder={placeholder}
-            width="100%"
-            sx={{ ...styles }}
-            onFocus={() => {
-              if (onFocus) {
-                onFocus();
-              }
-              if (openOnFocus) {
-                openMenu();
-              }
-            }}
-            onBlur={() => {
-              if (onBlur) {
-                onBlur();
-              }
-            }}
-            {...getInputProps()}
-          />
-          {/* right 9, because container box doesn't have border 1 */}
-          <Box position="absolute" right={9} lineHeight="normal">
-            <IconChevronStepper size={16} />
-          </Box>
-        </Box>
-
-        <Card
-          as="ul"
-          elevation={3}
-          display={isOpen ? 'block' : 'none'}
-          position="absolute"
-          float="left"
-          top="100%"
-          zIndex="1"
-          left={0}
-          mt="xs"
-          width={width}
-          maxHeight={maxHeight}
-          p={0}
-          m={0}
-          overflowY="auto"
-          {...getMenuProps()}
-        >
-          {inputItems.length !== 0 ? (
-            inputItems.map((item, index) => (
-              <Box
-                as="li"
-                px="md"
-                py="xs"
-                _hover={{
-                  backgroundColor: 'blue01',
-                }}
-                cursor="pointer"
-                textAlign="left"
-                lineHeight="20px"
-                fontSize={14}
-                sx={highlightedIndex === index ? { backgroundColor: 'blue01' } : {}}
-                key={`${item}_${index}`}
-                {...getItemProps({ item, index })}
-              >
-                {itemRenderer ? itemRenderer(item) : itemToString ? itemToString(item) : item}
-              </Box>
-            ))
-          ) : (
-            <Box
-              as="li"
-              px="md"
-              py="xs"
-              color="grey06"
-              cursor="pointer"
-              textAlign="left"
-              fontSize={14}
-              lineHeight="20px"
-            >
-              No items.
-            </Box>
-          )}
-        </Card>
-      </Stack>
-    </Box>
-  );
-}
-
-export default InputSelect;
