--- packages/react/src/components/form/components/InputSelectSearch/index.ts
@@ -1,2 +1,0 @@
-export { default as InputSelectSearch } from './InputSelectSearch';
-export * from './InputSelectSearch';
